+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects.API.Contracts/Project/Tasks/ReadinessToOpenMeetingTask.cs
@@ -1,7 +1,13 @@
+using Dfe.ManageFreeSchoolProjects.API.Contracts.Common;
+
 namespace Dfe.ManageFreeSchoolProjects.API.Contracts.Project.Tasks;
 
 public class ReadinessToOpenMeetingTask
 {
+    public YesNo? AROMIsExpectedToHappen { get; set; }
+
+    public DateTime? ExpectedDateOfTheMeeting { get; set; }
+
     public TypeOfMeetingHeld TypeOfMeetingHeld { get; set; }
     
     public DateTime? DateOfTheMeeting { get; set; }