+++ Dfe.ManageFreeSchoolProjects/Dfe.ManageFreeSchoolProjects.API/UseCases/Project/Tasks/ReadinessToOpenMeeting/ROMTaskBuilder.cs
@@ -1,5 +1,6 @@
 using Dfe.ManageFreeSchoolProjects.API.Contracts.Project.Tasks;
 using Dfe.ManageFreeSchoolProjects.Data.Entities.Existing;
+using Dfe.ManageFreeSchoolProjects.Data.Migrations;
 
 namespace Dfe.ManageFreeSchoolProjects.API.UseCases.Project.Tasks.ReadinessToOpenMeeting;
 
@@ -12,6 +13,8 @@ public static class ROMTaskBuilder
 
         return new ReadinessToOpenMeetingTask
         {
+            AROMIsExpectedToHappen = milestones.AROMIsExpectedToHappen,
+            ExpectedDateOfTheMeeting = milestones.FsgPreOpeningMilestonesRomForecastDate,
             DateOfTheMeeting = milestones.FsgPreOpeningMilestonesRomActualDateOfCompletion,
             TypeOfMeetingHeld = EnumParsers.ParseTypeOfMeetingHeld(milestones.ROMTypeOfMeetingHeld),
             WhyMeetingWasNotHeld = milestones.ROMWhyAMeetingWasNotHeld,