+++ release-notes.md
@@ -3,6 +3,7 @@
 * 178172 - Build: Iterate funding agreement health check and submission
 * 179446 - Build: Bulk edit projects
 * 182793 - Build: Set values for calculated database fields
+* 181754 - Build: readiness to open meeting (ROM) forecast date
 
 ## 0.13.0
 * 147110 - Build: Readiness to open meeting (ROM)